linux: refactoring to correctly track the current_task
The LinuxMonitor plugin currently uses the approach described at
https://stackoverflow.com/questions/11961490/understanding-the-getting-of-task-struct-pointer-from-process-kernel-stack
to get the current task_struct and hence the PID of the process currently
running in the system. Unfortunately, this approach is only valid for kernel
versions < 4.x In kernel versions > 4.x the only way to get the current
task_struct is via the get_current function
(arch/$(ARCH)/include/asm/current.h). This required some major refactoring to
get this information to the LinuxMonitor plugin.

We introduced a new dynamic kernel module, s2e (similar to the Windows S2E
driver). When loaded the S2E module sends crucial information on kernel
addresses and data structures to the LinuxMonitor plugin, including the address
of the current task_struct pointer. With this information, the LinuxMonitor
plugin is able to correctly track the current PID in the kernel.

We can also send offset values to the LinuxMonitor plugin, which makes the
plugin architecture independent and more easily ported to kernels (previously
these offsets were hardcoded into the LinuxMonitor plugin itself).

This module has the added bonus that we can correctly check if the LinuxMonitor
plugin is enabled when we resume from a snapshot. Previously, the LinuxMonitor
plugin was checked only at kernel boot. However, when resuming from a snapshot
the LinuxMonitor plugin may be in a different state (i.e. enabled or disabled)
than that when the booted was originally booted.

Finally, we rearchitectured the S2E modifications so that they can be
completely disabled via a Kconfig option.